backend/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
opencv-python-headless==4.9.0.80
numpy==1.26.4
ultralytics==8.0.196
pillow==10.0.1
websockets==12.0
python-multipart==0.0.9
onnxruntime==1.22.1